Checking in the scaffolding for a settings view
