+++ Apps/Relisten/Layouts/ArtistCellNode.swift
@@ -9,7 +9,6 @@
 import UIKit
 import RelistenShared
 
-import LayoutKit
 import FaveButton
 
 import AsyncDisplayKit
