+++ Shared/ShowCellNode.swift
@@ -5,7 +5,6 @@
 
 import UIKit
 
-import LayoutKit
 import AXRatingView
 
 import AsyncDisplayKit
