+++ Shared/View Controllers/SettingsViewController.swift
@@ -1,0 +1,74 @@
+//
+//  SettingsViewController.swift
+//  RelistenShared
+//
+//  Created by Jacob Farkas on 8/26/18.
+//  Copyright © 2018 Alec Gorge. All rights reserved.
+//
+
+import Foundation
+import AsyncDisplayKit
+
+public class SettingsViewController : RelistenBaseAsyncTableViewController {
+    enum Sections: Int, RawRepresentable {
+        case downloads = 0
+        case bugReporting
+        case credits
+        case count
+    }
+}
+
+// MARK: ASTableDataSource
+extension SettingsViewController {
+    func numberOfSections(in tableNode: ASTableNode) -> Int {
+        return Sections.count.rawValue
+    }
+    
+    func tableNode(_ tableNode: ASTableNode, numberOfRowsInSection section: Int) -> Int {
+        switch Sections(rawValue: section)! {
+        case .downloads:
+            fallthrough
+        case .bugReporting:
+            fallthrough
+        case .credits:
+            return 1
+        case .count:
+            fatalError()
+        }
+    }
+    
+    public func tableNode(_ tableNode: ASTableNode, nodeForRowAt indexPath: IndexPath) -> ASCellNode {
+        var n: ASCellNode
+        
+        switch Sections(rawValue: indexPath.section)! {
+        case .downloads:
+            fallthrough
+        case .bugReporting:
+            fallthrough
+        case .credits:
+            fallthrough
+            
+        case .count:
+            fatalError()
+        }
+        
+        return n
+    }
+    
+    func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
+        switch Sections(rawValue: section)! {
+        case .downloads:
+            return "Offline Music"
+        case .bugReporting:
+            return "Bug Reporting"
+        case .credits:
+            return "Credits"
+        case .count:
+            fatalError()
+        }
+    }
+    
+    func tableView(_ tableView: UITableView, shouldHighlightRowAt indexPath: IndexPath) -> Bool {
+        return false
+    }
+}
